패킷 사이즈 ushort 캐스팅 누락부분 추가

+++ Server/ClientSession.cs
@@ -24,7 +24,7 @@ namespace Server
             
             byte[] sendBuffer = new byte[size + 4];
             // write size
-            Array.Copy(BitConverter.GetBytes(size + 4), 0, sendBuffer, 0, sizeof(ushort));
+            Array.Copy(BitConverter.GetBytes((ushort)(size + 4)), 0, sendBuffer, 0, sizeof(ushort));
             // write id
             Array.Copy(BitConverter.GetBytes((ushort)msgId), 0, sendBuffer, 2, sizeof(ushort));
             // write encoded
